[Refactor] Marker 클래스 수정
