+++ Assets/Scripts/SeongJae/MarkerModel.cs
@@ -1,8 +1,6 @@
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 
-public class MarkerModel : MonoBehaviour
+public class MarkerModel
 {
 
 }
